@@ -28,7 +28,7 @@ import { IconSize } from '../../icons/icons.enum';
 export class AvatarComponent implements OnChanges, OnInit, AfterViewInit {
   constructor(private host: ElementRef, private DOM: DOMhelpers) {}
 
-  @ViewChild('content') private content: ElementRef;
+  @ViewChild('content', { static: false }) private content: ElementRef;
 
   public hasContent = true;
   readonly avatarSize = AvatarSize;
