@@ -19,9 +19,9 @@ import { PanelPositionService } from '../../popups/panel/panel-position-service/
 import { BaseFormElement } from '../base-form-element';
 
 export abstract class BaseSelectPanelElement extends BaseFormElement {
-  @ViewChild(CdkOverlayOrigin) overlayOrigin: CdkOverlayOrigin;
-  @ViewChild('templateRef') templateRef: TemplateRef<any>;
-  @ViewChild('triggerInput', { read: ElementRef }) triggerInput: ElementRef;
+  @ViewChild(CdkOverlayOrigin, { static: true }) overlayOrigin: CdkOverlayOrigin;
+  @ViewChild('templateRef', { static: true }) templateRef: TemplateRef<any>;
+  @ViewChild('triggerInput', { read: ElementRef, static: true }) triggerInput: ElementRef;
 
   @Input() isQuickFilter = false;
 
