@@ -39,8 +39,8 @@ import { animate, style, transition, trigger } from '@angular/animations';
   ],
 })
 export class QuickFilterBarComponent implements OnChanges, AfterViewInit {
-  @ViewChild('prefix') prefix: ElementRef;
-  @ViewChild('suffix') suffix: ElementRef;
+  @ViewChild('prefix', { static: false }) prefix: ElementRef;
+  @ViewChild('suffix', { static: false }) suffix: ElementRef;
 
   @Input() quickFilters: QuickFilterConfig[];
   @Input() showResetFilter = false;
