@@ -37,10 +37,10 @@ export class MockComponent implements AfterViewInit {
   @Input() slot3css = {};
   @Input() slot4css = {};
 
-  @ViewChild('slot1') slot1: ElementRef;
-  @ViewChild('slot2') slot2: ElementRef;
-  @ViewChild('slot3') slot3: ElementRef;
-  @ViewChild('slot4') slot4: ElementRef;
+  @ViewChild('slot1', { static: false }) slot1: ElementRef;
+  @ViewChild('slot2', { static: false }) slot2: ElementRef;
+  @ViewChild('slot3', { static: false }) slot3: ElementRef;
+  @ViewChild('slot4', { static: false }) slot4: ElementRef;
 
   hasSlots = [true, true, true, true];
 
